refactor: extract dummy movie data into reusable test data file
